+++ CHANGELOG.md
@@ -18,6 +18,7 @@ form of `process_messages_{a,b,e}` is now `TextProduct,str`.
 - Add option to `mcalc_feelslike` to support `mask_undefined`.
 - Add `twitter_media` link for generic text products that have a polygon (#586).
 - Add `limit_by_doy` option to `windrose_utils` to allow a day of year limit.
+- Add parser for SPC Watch Probabilities (WWP) product (#595).
 - Allow `pyiem.nws.nwsli` instance to be subscriptable for iterop.
 - Support passing `linewidths` to `MapPlot.contourf`.